refactor(registry): split organizations.rs into sub-modules
Split src/registry/organizations.rs (525 lines) into 9 focused
sub-modules under organizations/, each under 100 lines, to comply
with the project's file size limit.

Module breakdown:
- mod.rs (18 lines) - declarations and re-exports
- errors.rs (32 lines) - OrganizationError enum
- slug.rs (70 lines) - slugify/validate_slug + tests
- query.rs (58 lines) - list/get organizations
- create.rs (54 lines) - create organization
- update.rs (96 lines) - update organization
- delete.rs (34 lines) - delete organization
- assignment.rs (100 lines) - set project organization
- org_file.rs (38 lines) - read/write .centy/organization.json
- sync.rs (62 lines) - sync org from project file

Closes #124
